requirements.txt
annotated-types==0.7.0
blinker==1.8.2
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fonttools==4.53.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
matplotlib==3.9.0
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
preshed==3.0.9
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pyodbc==5.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.5.15
requests==2.32.2
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-lookups-data==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
thinc==8.2.3
threadpoolctl==3.5.0
tqdm==4.66.4
typer==0.9.4
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.3
zipp==3.19.0